combined AlchemyLanguage and AlchemyData News into one service and data model
