--- Scripts/Services/AlchemyDataNews.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 4d6ca99c6a8101841bd491ca62f1b13f
-folderAsset: yes
-timeCreated: 1467221141
-licenseType: Pro
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
