--- Scripts/Services/AlchemyDataNews/AlchemyDataNews.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 05a500ab2e1dbe845a759f967f933e9d
-timeCreated: 1467221163
-licenseType: Pro
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
