--- Scripts/Services/AlchemyDataNews/DataModels.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 8a5d52bfdcd1aa54a88cb39c632fc626
-timeCreated: 1467221148
-licenseType: Pro
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
